Seperate BitcoinDatabase from DAo
